Experiment in rendering links to all artifacts.
This experiment revealed it is problematic to render
all artifacts on a single page because of how the output templates
are tightly coupled to modules. In fact, it is not even possible to
link to the artifact until the module itself is started.

This is because modules are questions from the perspective
of the project’s root task (e.g., `app.yaml`). In the legacy code,
when a module is started, task records are created in the database
paired with the questions. The artifact needs access to the task
records and the task answer records to render.

A possible interim solution is to move all artifacts up to the top level, or
root task of a project (e.g., `app.yaml`). Moving all artifacts to the top
level, to the root_task level would make it easy to render the artifacts. In
fact, the code already exists to do show documents that exist at the top
level.